Split cli.py and tests.py into several modules
