+++ .gitignore
@@ -5,3 +5,4 @@ tmp/
 .tox/
 .coverage
 __pycache__/
+tags
