+++ .travis.yml
@@ -7,7 +7,7 @@ install:
     - pip install pytest coverage coveralls flake8
     - pip install -e .
 script:
-    - coverage run -m pytest tests.py
+    - coverage run -m pytest tests
     - coverage report -m --fail-under=100
     - flake8 *.py
 after_script:
