+++ src/check_python_versions/parsers/__init__.py
@@ -1,0 +1,1 @@
+# make a package
