+++ src/check_python_versions/parsers/travis.py
@@ -1,0 +1,45 @@
+try:
+    import yaml
+except ImportError:  # pragma: nocover
+    yaml = None
+
+from .tox import parse_envlist, tox_env_to_py_version
+
+
+TRAVIS_YML = '.travis.yml'
+
+
+def get_travis_yml_python_versions(filename=TRAVIS_YML):
+    with open(filename) as fp:
+        conf = yaml.safe_load(fp)
+    versions = []
+    if 'python' in conf:
+        versions += map(travis_normalize_py_version, conf['python'])
+    if 'matrix' in conf and 'include' in conf['matrix']:
+        for job in conf['matrix']['include']:
+            if 'python' in job:
+                versions.append(travis_normalize_py_version(job['python']))
+    if 'jobs' in conf and 'include' in conf['jobs']:
+        for job in conf['jobs']['include']:
+            if 'python' in job:
+                versions.append(travis_normalize_py_version(job['python']))
+    if 'env' in conf:
+        toxenvs = []
+        for env in conf['env']:
+            if env.startswith('TOXENV='):
+                toxenvs.extend(parse_envlist(env.partition('=')[-1]))
+        versions.extend(
+            tox_env_to_py_version(e) for e in toxenvs if e.startswith('py'))
+    return sorted(set(versions))
+
+
+def travis_normalize_py_version(v):
+    v = str(v)
+    if v.startswith('pypy3'):
+        # could be pypy3, pypy3.5, pypy3.5-5.10.0
+        return 'PyPy3'
+    elif v.startswith('pypy'):
+        # could be pypy, pypy2, pypy2.7, pypy2.7-5.10.0
+        return 'PyPy'
+    else:
+        return v
