+++ tests/test_utils.py
@@ -1,0 +1,5 @@
+from check_python_versions.utils import pipe
+
+
+def test_pipe():
+    assert pipe('echo', 'hi') == 'hi\n'
