+++ tox.ini
@@ -4,18 +4,19 @@ envlist = flake8,py36,py37
 [testenv]
 deps = pytest
 commands =
-    pytest tests.py {posargs}
+    pytest {posargs:tests}
 
 [testenv:pypy36]
 basepython = pypy3.6
 
 [testenv:coverage]
+usedevelop = true
 basepython = python3.6
 deps =
     {[testenv]deps}
     coverage
 commands =
-    coverage run -m pytest tests.py {posargs}
+    coverage run -m pytest tests {posargs}
 ##  coverage report -m --fail-under=100
     coverage report -m
 
@@ -24,4 +25,4 @@ basepython = python3.6
 deps =
     flake8
 commands =
-    flake8 *.py
+    flake8 src *.py
